requirements.txt
# CAI-System Dependencies
# Core data processing
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# I/O and utilities
tqdm>=4.65.0
requests>=2.28.0
pyyaml>=6.0

# NLP for object matching
nltk>=3.8.0

# Visualization
matplotlib>=3.7.0
opencv-python>=4.8.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (optional)
# black>=23.0.0
# flake8>=6.0.0
